+++ src/app/shared/services/app-auth-guard/index.ts
@@ -1,0 +1,2 @@
+export * from './app-auth-guard.service';
+export * from './roles';
