+++ src/app/shared/services/index.ts
@@ -1,2 +1,2 @@
 export * from './query-params-store';
-export * from './app-auth-guard.service';
+export * from './app-auth-guard';
